+++ src/auth/ValClient/Account.ts
@@ -2,8 +2,10 @@
 import { CookieJar as toughCookie } from "tough-cookie";
 
 import * as IngCore from "@ing3kth/core";
-import type { IValClient_Auth } from "../../resources/interface/IValClient";
 import { AuthFlow } from "./AuthFlow";
+
+import type { IValClient_Auth } from "../../resources/interface/IValClient";
+import type { IAxiosClient_Out } from "@ing3kth/core/dist/interface/IAxiosClient";
 
 //class
 
@@ -11,18 +13,18 @@ import { AuthFlow } from "./AuthFlow";
  * * Class ID: @ing3kth/val-api/Account
  */
 class Account {
-    classId:string;
-    cookie:toughCookie;
-    accessToken:string;
-    id_token:string;
-    expires_in:number;
-    token_type:string;
-    entitlements:string;
-    region: {
+    public classId:string;
+    private cookie:toughCookie;
+    private accessToken:string;
+    private id_token:string;
+    private expires_in:number;
+    private token_type:string;
+    private entitlements:string;
+    private region: {
         pbe: string,
         live: string,
     };
-    multifactor:boolean;
+    public multifactor:boolean;
 
     constructor() {
         this.classId = '@ing3kth/val-api/Account';
@@ -44,8 +46,8 @@ class Account {
      * @param {String} password Riot Account Password
      * @returns {Promise<IValClient_Auth>}
      */
-    async execute(username:string, password:string):Promise<IValClient_Auth> {
-        const axiosClient = new IngCore.AxiosClient({
+    public async execute(username:string, password:string):Promise<IValClient_Auth> {
+        const axiosClient:IngCore.AxiosClient = new IngCore.AxiosClient({
             cookie: true,
             jar: this.cookie.toJSON(),
             headers: {}
@@ -65,7 +67,7 @@ class Account {
         });
 
         //ACCESS TOKEN
-        const auth_response = await axiosClient.put('https://auth.riotgames.com/api/v1/authorization', {
+        const auth_response:IAxiosClient_Out = await axiosClient.put('https://auth.riotgames.com/api/v1/authorization', {
             'type': 'auth',
             'username': username,
             'password': password,
@@ -84,7 +86,7 @@ class Account {
      * 
      * @returns {IValClient_Auth}
      */
-    toJSON():IValClient_Auth {
+    public toJSON():IValClient_Auth {
         IngCore.Logs.log("Export " + this.classId);
 
         return {
@@ -105,7 +107,7 @@ class Account {
      * @returns {Promise<IValClient_Auth>}
      */
     static async login(username:string, password:string):Promise<IValClient_Auth> {
-        const NewAccount = new Account();
+        const NewAccount:Account = new Account();
         return await NewAccount.execute(username, password);
     }
 }